+++ core/src/cljfmt/task/core.clj
@@ -193,7 +193,7 @@
   "Check a single source file and produce a result."
   [config path ^File file]
   (let [original (slurp file)
-        revised (format/reformat-string original config)]
+        revised (format/reformat-file original config)]
     (if (= original revised)
       {:type :correct
        :debug (str "Source file " path " is formatted correctly")}
@@ -236,7 +236,7 @@
   "Fix a single source file and produce a result."
   [config path ^File file]
   (let [original (slurp file)
-        revised (format/reformat-string original config)]
+        revised (format/reformat-file original config)]
     (if (= original revised)
       {:type :correct
        :debug (str "Source file " path " is formatted correctly")}