+++ util/completion.zsh
@@ -1,0 +1,18 @@
+#compdef cljstyle
+
+# ZSH completion for cljstyle.
+# Drop this somewhere on your `$fpath` named `_cljstyle`.
+
+_cljstyle() {
+    _arguments \
+        "--stats=[statistics file]:Write formatting stats to this file:_files" \
+        "--no-color[disable color output]" \
+        "--verbose[verbose logging output]" \
+        "--help[show usage information]" \
+        "1:command:((find\\:'Find files which would be processed'
+                     check\\:'Check source files and print a diff for errors'
+                     fix\\:'Edit source files to fix formatting errors'
+                     config\\:'Show config used for a given path'
+                     version\\:'Print program version information'))" \
+        "*::path:_files"
+}